fix cluster stuff
